feat: move guard to separate module
