+++ src/app.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get } from '@nestjs/common';
 import { AppService } from './app.service';
-import { Public } from './auth/auth.decorator';
+import { Public } from './security/security.decorator';
 
 @Controller()
 export class AppController {
