+++ src/security/security.guard.ts
@@ -7,7 +7,7 @@ import {
 import { Reflector } from '@nestjs/core';
 import { JwtService } from '@nestjs/jwt';
 import { Request } from 'express';
-import { IS_PUBLIC_KEY } from './auth.decorator';
+import { IS_PUBLIC_KEY } from './security.decorator';
 import { ConfigService } from '@nestjs/config';
 
 @Injectable()
